requirements_test.txt
# Test requirements
pillow>=7
pytest-django>=3.5.1,<3.6
pytest>=5.0,<5.1
pytest-cov>=2.7.1
flake8==3.7.8

djangorestframework>=2.4.3
